% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-vector.R
\name{arrivals_state_vector}
\alias{arrivals_state_vector}
\title{Get state vectors for arrivals at airport}
\usage{
arrivals_state_vector(session, apt, wef, til = NULL, duration = 3600)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{apt}{ICAO ID of airport, i.e. "EDDF" for Frankfurt}

\item{wef}{(UTC) timestamp of With Effect From (included)}

\item{til}{(UTC) timestamp of TILl instant (excluded), if NULL
if is interpreted as WEF + 1 day.}

\item{duration}{number of second back from `lastseen`}
}
\value{
data frame of state vector data of flights containing the following
        variables (see also OSN docs about
  \href{https://opensky-network.org/apidoc/rest.html#arrivals-by-airport}{Arrivals
   by Airport}):
   \tabular{lll}{
     \strong{Name}       \tab \strong{Description} \tab \strong{Type} \cr
     icao24              \tab ICAO 24-bit address \tab chr \cr
     callsign            \tab flight's callsign   \tab chr \cr
     estdepartureairport \tab Estimated departure airport \tab chr \cr
     estarrivalairport   \tab Estimated arrival airport   \tab chr \cr
     start               \tab Start of portion of trajectory: `min(firstseen, lastseen - duration)` \tab int \cr
     firstseen           \tab first seen by OpenSky Network (UNIX timestamp)\tab int \cr
     lastseen            \tab last seen by OpenSky Network (UNIX timestamp) \tab int \cr
     item.time           \tab position report's time (UNIX timestamp) \tab int \cr
     item.longitude      \tab position report's longitude (WSG84 decimal degrees)\tab dbl \cr
     item.latitude       \tab position report's latitude (WSG84 decimal degrees) \tab dbl \cr
     item.velocity       \tab ground speed of the aircraft (m/s) \tab dbl \cr
     item.heading        \tab true track in decimal degrees clockwise from north (north=0°) \tab dbl \cr
     item.vertrate       \tab vertical speed of the aircraft (m/s) \tab dbl \cr
     item.onground       \tab TRUE if the position was retrieved from a surface position report \tab lgl \cr
     item.baroaltitude   \tab position report's barometric altitude (m) \tab dbl \cr
     item.geoaltitude    \tab position report's GNSS (GPS) altitude (m) \tab dbl \cr
     item.hour           \tab position report's hour (UNIX timestamp) \tab int
   }
}
\description{
Retrive state vectors for all flights from flights table
}
\examples{
\dontrun{
session <- connect_osn("cucu", verbose = 2)
arrivals_state_vector(session, "EDDF", "2019-04-22 00:00:00", til=NULL)
}
}
