% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-vector.R
\name{minimal_state_vector}
\alias{minimal_state_vector}
\title{Get few state vector's attributes for a time period (listing who was present)}
\usage{
minimal_state_vector(session, icao24, wef, til = NULL, bbox = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{icao24}{(Optional) Single or vector of ICAO24 ICAO 24-bit addresses}

\item{wef}{Start of period of interest (date or datetime)}

\item{til}{(Optional) End of period of interest, if NULL wef_time + 1 day}

\item{bbox}{(Optional) axis aligned bounding box like
`c(xmin, xmax, ymin, ymax)`}
}
\value{
a dataframe of state vectors with distinct icao24, callsign and hour
}
\description{
Get few state vector's attributes for a time period (listing who was present)
}
\examples{
\dontrun{
session <- connect_osn("cucu", verbose = 2)
minimal_state_vector(
   session,
   icao24 = c("3c6589", "3c6757"),
   wef = "2019-04-22 00:00:00",
   til = "2019-04-22 10:00:00",
   bbox = c(xmin = 7.536746, xmax = 9.604390, ymin = 49.36732, ymax = 50.69920)
)
}
}
