% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational-status.R
\name{operational_status}
\alias{operational_status}
\title{Get operational status for a time period}
\usage{
operational_status(session, icao24, wef_time, til_time = NULL,
  bbox = NULL, debug_level = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{icao24}{(Optional) Single or vector of ICAO24 ICAO 24-bit addresses
among those in the bounding box}

\item{wef_time}{Start of period of interest}

\item{til_time}{(Optional) End of period of interest, if NULL wef_time + 1 day}

\item{bbox}{(Optional) axis aligned bounding box like
`c(xmin, xmax, ymin, ymax)`}

\item{debug_level}{one of "DEBUG", "TRACE" or NULL (default)
It prints usefult debug info, i.e. input args, query string,
or traces the query results in the file `query_output.txt`.}
}
\value{
a dataframe of operational status info (icao24, hour, rawmsg)
}
\description{
Get operational status for a time period
}
\examples{
\dontrun{
osn:::operational_status(
  session = 1,
  icao24 = NULL,
  wef_time = "2019-01-01 00:00:00",
  til_time = "2019-01-01 02:00:00",
  bbox = c(xmin = 7.536746, xmax = 9.604390, ymin = 49.36732, ymax = 50.69920),
  debug_level = "DEBUG"
)
}
}
