% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{osn_connect}
\alias{osn_connect}
\title{Create an ssh session to OpenSky Network’s Impala shell.}
\usage{
osn_connect(
  usr,
  passwd = askpass,
  host = "data.opensky-network.org",
  port = 2230,
  verbose = FALSE
)
}
\arguments{
\item{usr}{user account}

\item{passwd}{either a string or a callback function for password prompt}

\item{host}{an ssh server string of the form \code{[user@]hostname[:@port]}. An ipv6
hostname should be wrapped in brackets like this: \code{[2001:db8::1]:80}.}

\item{port}{port to connect to}

\item{verbose}{either TRUE/FALSE or a value between 0 and 4 indicating log level:
0: no logging, 1: only warnings, 2: protocol, 3: packets or 4: full stack trace.}
}
\value{
an SSH session
}
\description{
Create an ssh session to OpenSky Network’s Impala shell.
}
\examples{
\dontrun{
# connect directly to OSN
session <- osn_connect("cucu", verbose = 2)

# connect via SSH port forwarding
session <- osn_connect_osn(
  usr = Sys.getenv("OSN_USER"),
  passwd = Sys.getenv("OSN_PASSWORD"),
  port = 6666,
  host = "localhost"
)
}
}
