% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{osn_disconnect}
\alias{osn_disconnect}
\title{Disconnect from OpenSky Network’s Impala shell.}
\usage{
osn_disconnect(session)
}
\arguments{
\item{session}{ssh connection created with \code{\link[ssh:ssh_connect]{ssh_connect()}}}
}
\value{
an SSH session
}
\description{
Disconnect from OpenSky Network’s Impala shell.
}
\examples{
\dontrun{
session <- osn_connect("cucu", verbose = 2)
osn_disconnect(session)
}
}
