% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-vector.R
\name{minimal_state_vector}
\alias{minimal_state_vector}
\title{Get few state vector's attributes for a time period (listing who was present)}
\usage{
minimal_state_vector(session, icao24, wef_time, til_time = NULL,
  bbox = NULL, debug_level = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{icao24}{(Optional) Single or vector of ICAO24 ICAO 24-bit addresses}

\item{wef_time}{Start of period of interest}

\item{til_time}{(Optional) End of period of interest, if NULL wef_time + 1 day}

\item{bbox}{(Optional) axis aligned bounding box like
`c(xmin, xmax, ymin, ymax)`}

\item{debug_level}{one of "DEBUG", "TRACE" or NULL (default)
It prints usefult debug info, i.e. input args, query string,
or traces the query results in the file `query_output.txt`.}
}
\value{
a dataframe of state vectors with distinct icao24, callsign and hour
}
\description{
Get few state vector's attributes for a time period (listing who was present)
}
\examples{
\dontrun{
session <- connect_osn("cucu", verbose = 2)
minimal_state_vector(
   session,
   icao24 = c("3c6589", "3c6757"),
   wef_time = "2019-04-22 00:00:00",
   til_time = "2019-04-22 10:00:00"
)
}
}
