% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-vector.R
\name{state_vector}
\alias{state_vector}
\title{Get state vectors for a time period}
\usage{
state_vector(session, icao24, wef_time, til_time = NULL, bbox = NULL)
}
\arguments{
\item{session}{SSH session to OSN Impala}

\item{icao24}{(Optional) Single or vector of ICAO24 ICAO 24-bit addresses}

\item{wef_time}{Start of period of interest}

\item{til_time}{(Optional) End of period of interest, if NULL wef_time + 1 day}

\item{bbox}{(Optional) axis aligned bounding box
(lon_min, lat_min, lon_max, lat_max)}
}
\value{
a dataframe of state vectors
}
\description{
Get state vectors for a time period
}
\examples{
\dontrun{
session <- connect_osn("cucu", verbose = 2)
state_vector(
   session,
   icao24 = c("3c6589", "3c6757"),
   wef_time = "2019-04-22 00:00:00",
   til_time = "2019-04-22 10:00:00"
)
}
}
